workspace/backend/requirements.txt
alembic==1.4.0
aniso8601==8.0.0
astroid==2.3.3
Babel==2.8.0
Click==7.0
coverage==5.0.3
# elliptic curve digital signature algorithm library
ecdsa==0.15
factory-boy==2.12.0
# python 2 backward compatibility
future==0.18.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
# python import sort:
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
# lazy object proxy:
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
# code checker:
mccabe==0.6.1
pylint==2.3.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
# A JOSE implementation in Python using pycryptodome instead pycrypto
python-jose-cryptodome==1.3.2
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
# the standard Python AST (plus type comments) generator
typed-ast==1.4.1
uWSGI==2.0.18
Werkzeug==0.16.0
WTForms==2.2.1